23.04.09 스터디 매니저 페이지 int -> Long 처리 조회수나 게시글 수가 0일 때 int이면 오류가 뜸, Long으로 처리해서 오류 안 뜨게 수정

+++ src/main/java/highsquare/hirecoder/web/controller/StudyManageController.java
@@ -44,9 +44,9 @@ public class StudyManageController {
     public String getManagerMainPage(@PathVariable("studyId") Long studyId,
                                      Model model) {
         // 해당 스터디의 게시글 수
-        int boardCount = dashBoardService.getCountBoard(studyId, CONTENT);
+        Long boardCount = dashBoardService.getCountBoard(studyId, CONTENT);
         // 해당 스터디 글의 전체 조회수
-        int countViewCnt = dashBoardService.getViewCntBoard(studyId, CONTENT);
+        Long countViewCnt = dashBoardService.getViewCntBoard(studyId, CONTENT);
         // 해당 스터디의 스터디원 수
         // 여기서 Study를 불러와서 변수에 다 저장하는 게 좋을까??
         // 이 과정을 컨트롤러에서 해도 되는지 모르겠음
